resolving comments
