+++ src/features/mod.rs
@@ -1,0 +1,2 @@
+// All features should be registered here
+pub mod cleaner;
