+++ src/main.rs
@@ -1,8 +1,7 @@
-use std::fs;
-use anyhow::{Context, Result};
-use walkdir::WalkDir;
+use anyhow::Result;
 
 mod arg;
+mod features;
 
 fn main() -> Result<()> {
     let args = arg::parse_args();
@@ -11,41 +10,12 @@ fn main() -> Result<()> {
     println!("File types to clean: {:?}", args.types);
     println!("Minimum file size: {} bytes", args.min_size);
 
-    directory_cleaner(&args.dir, &args.types, args.min_size)?;
+    features::cleaner::directory_cleaner(&args.dir, &args.types, args.min_size)?;
 
     println!("Cleaning completed successfully.");
 
     Ok(())
 }
-
-
-fn directory_cleaner(dir: &String, paths_to_clear: &[String], size: u64) -> Result<()> {
-    // skip files that cannot be opened, due to permission issues etc
-    for entry in WalkDir::new(dir).into_iter().filter_map(|f| f.ok()) {
-        let path = entry.path();
-
-        if path.is_file() {
-            let ext = path.extension()
-                .and_then(|ex| ex.to_str())
-                .unwrap_or("");
-
-            if paths_to_clear.iter().any(|p| ext == p) {
-                let metadata = fs::metadata(path)
-                    .with_context(|| format!("Failed to read metadata for file: {:?}", path))?;
-
-                if metadata.len() >= size {
-                    println!("Deleting file: {:?}", path);
-                    fs::remove_file(path)
-                        .with_context(|| format!("Failed to delete file: {:?}", path))?;
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-
 
 #[cfg(test)]
 mod tests {
@@ -62,7 +32,7 @@ mod tests {
 
         let file_types = vec!["txt".to_string()];
         let dir_str = temp_dir.path().to_str().unwrap().to_string();
-        directory_cleaner(&dir_str, &file_types, 500)?;
+        features::cleaner::directory_cleaner(&dir_str, &file_types, 500)?;
 
         assert!(!file_path.exists(), "File should have been deleted");
 
@@ -78,9 +48,12 @@ mod tests {
 
         let file_types = vec!["txt".to_string()];
         let dir_str = temp_dir.path().to_str().unwrap().to_string();
-        directory_cleaner(&dir_str, &file_types, 2000)?;
+        features::cleaner::directory_cleaner(&dir_str, &file_types, 2000)?;
 
-        assert!(file_path.exists(), "File should not have been deleted due to size constraint");
+        assert!(
+            file_path.exists(),
+            "File should not have been deleted due to size constraint"
+        );
 
         // temp_dir will be cleaned up here when it goes out of scope
